Introduced BaseWebSecurityConfig to be able to provide different security configurations in production and testing scenarios. Default implementations for a few methods are provided. Subclasses are WebSecurityConfig and TestWebSecurityConfig. Additionally, OaspProfile is introduced which provides String constants to distinguish between production and test scenarios. Also, a RestTestClientBuilder is introduced which provides a static method to construct JAXRSClients. Please, refer to TablemanagementRestServiceTest for exmaples of how to used the added classes
